Add URL busting for CSS files
The mechanism is relatively simple: /static/ANYTHING/path/to/file will return
the file. Thus, to bust the cache, we just have to change what we are putting in
ANYTHING. To do that, we just look at the modification time of the CSS file and
use that (which will probably be anytime a deploy happens). Importantly, the
point here is not to ever be able to serve an old version, just to get browsers
to re-fetch when necessary (so whenever a deploy happens, the mtime changes, and
thus the url served will change). We could also probably just use the startup
time for the application, or something else that'll change in some reasonable
way.