Make files stored in S3 be encrypted with public key before being stored.
The private key is put into the session by a curator (not stored anywhere else),
which is used to decrypt and show them files.

To make it easier to get this going, this is using the crypto-simple wrapper
library around cryptonite, but that requires strict bytestrings which _may_ end
up being a problem as it means we need to load files into memory in order to
encrypt/decrypt them. but, we'll see if this actually matters.