Improve amount UI, change summer guarantee language to summer typical
The latter because no one except 1&2 year CS grads have guaranteed summer
funding, so this will skew to the point of meaninglessness the stipends. What is
important is whether the stipend is typically experienced to be year round,
rather than what is "contractually" guaranteed (given that we have no contract).